fix error with repeats
